requirements.txt
aiogram==2.25.1
qrcode
python-dotenv==1.0.0